+++ src/config.c
@@ -312,6 +312,7 @@ const config default_cfg = {
   old_options : "",
   dim_margins : false,
   status_line : false,
+  status_debug : 0,
   old_xbuttons : false,
   use_system_colours : false,
   old_bold : false
@@ -609,6 +610,7 @@ options[] = {
   {"Bloom", OPT_INT, offcfg(bloom)},
   {"DimMargins", OPT_BOOL, offcfg(dim_margins)},
   {"StatusLine", OPT_BOOL, offcfg(status_line)},
+  {"StatusDebug", OPT_INT, offcfg(status_debug)},
   {"OldXButtons", OPT_BOOL, offcfg(old_xbuttons)},
   {"OptionsFont", OPT_WSTRING, offcfg(options_font)},
   {"OptionsFontSize", OPT_INT | OPT_LEGACY, offcfg(options_fontsize)},
@@ -2815,7 +2817,65 @@ download_scheme(char * url)
     }
   }
   else {
+    int l = 0;
+    char linebuf[22222];  // in case of json, pull in the whole stuff
     while (fgets(linebuf, sizeof(linebuf) - 1, sf)) {
+#if defined(debug_scheme) && debug_scheme > 1
+      printf("linebuf <%s>\n", linebuf);
+#endif
+
+      if (!l++ && *linebuf == '{') {
+        // handle drag-and-drop json formats that contain colour specs like 
+        // "Red=190,70,120" (https://github.com/mskyaxl/wsl-terminal) or
+        // "Red=220,50,47\r" (https://github.com/oumu/mintty-color-schemes)
+        auto schapp = [&](const char * name)
+        {
+          char specbuf[30];
+          sprintf(specbuf, "\"%s=", name);
+          char * colspec = strstr(linebuf, specbuf);
+          if (!colspec)
+            return;
+          colspec++;
+          char * cpoi = colspec + strlen(name) + 1;
+          while (isdigit((uchar)*cpoi) || *cpoi == ',')
+            cpoi++;
+          int collen = cpoi - colspec;
+          int len = sch ? strlen(sch) : 0;
+          sch = renewn(sch, len + collen + 2);
+          snprintf(&sch[len], collen + 1, "%s", colspec);
+          sprintf(&sch[len + collen], ";");
+#if defined(debug_scheme) && debug_scheme > 1
+          printf("%s\n", &sch[len]);
+#endif
+        };
+        schapp("ForegroundColour");
+        schapp("BackgroundColour");
+        schapp("BoldColour");
+        schapp("BlinkColour");
+        schapp("CursorColour");
+        schapp("UnderlineColour");
+        schapp("HoverColour");
+        schapp("HighlightBackgroundColour");
+        schapp("HighlightForegroundColour");
+        schapp("Black");
+        schapp("Red");
+        schapp("Green");
+        schapp("Yellow");
+        schapp("Blue");
+        schapp("Magenta");
+        schapp("Cyan");
+        schapp("White");
+        schapp("BoldBlack");
+        schapp("BoldRed");
+        schapp("BoldGreen");
+        schapp("BoldYellow");
+        schapp("BoldBlue");
+        schapp("BoldMagenta");
+        schapp("BoldCyan");
+        schapp("BoldWhite");
+        goto scheme_return;
+      }
+
       char * eq = linebuf;
       while ((eq = strchr(++eq, '='))) {
         int dum;
@@ -2857,6 +2917,9 @@ download_scheme(char * url)
       }
     }
   }
+
+scheme_return:
+
 #ifdef use_curl
   pclose(sf);
 #else