+++ src/winxx.cc
@@ -222,7 +222,7 @@ extern "C" {
       Tab& tab = tabs.back();
       tab.terminal->child = tab.chld.get();
       tab.terminal.get()->show_scrollbar = !!cfg.scrollbar;  // hotfix #597
-      (term_resize)(tab.terminal.get(), rows, cols);
+      (term_resize)(tab.terminal.get(), rows, cols, false);
       (term_reset)(tab.terminal.get(), true);
       tab.chld->cmd = g_cmd;
       tab.chld->home = g_home;